+++ tycho-metadata-model/src/test/java/org/eclipse/tycho/maven/test/EclipseModelTest.java
@@ -16,13 +16,11 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
 import org.eclipse.tycho.model.Feature;
 import org.eclipse.tycho.model.FeatureRef;
-import org.eclipse.tycho.model.Platform;
 import org.eclipse.tycho.model.PluginRef;
 import org.eclipse.tycho.model.UpdateSite;
 import org.junit.jupiter.api.BeforeEach;
@@ -117,57 +115,6 @@ class EclipseModelTest {
         assertEquals("http://www.example.com/license_MODIFIED", feature.getLicenseURL());
         assertEquals("http://www.example.com/copyright", feature.getCopyrightURL());
         assertEquals("Test Copyright", feature.getCopyright());
-    }
-
-    @Test
-    void testPlatform() throws Exception {
-        Platform platform = Platform.read(new File("src/test/resources/modelio/platform.xml"));
-
-        assertEquals(false, platform.isTransient());
-
-        List<Platform.Site> sites = platform.getSites();
-        assertEquals(2, sites.size());
-
-        List<String> plugins = sites.get(0).getPlugins();
-        assertEquals(2, plugins.size());
-        assertEquals("m2eclipse/org.maven.ide.components.archetype-common/", plugins.get(0));
-
-        List<Platform.Feature> features = sites.get(1).getFeatures();
-        assertEquals(2, features.size());
-
-        Platform transientPlatform = new Platform(platform);
-        transientPlatform.setTransient(true);
-
-        Platform.Site transientSite = new Platform.Site("file:/xxx");
-        transientPlatform.addSite(transientSite);
-
-        List<String> transientPlugins = new ArrayList<>();
-        transientPlugins.add("plugins/yyy/");
-        transientPlugins.add("plugins/zzz/");
-        transientSite.setPlugins(transientPlugins);
-
-        List<Platform.Feature> transientFeatures = new ArrayList<>();
-
-        Platform.Feature transientFeature = new Platform.Feature();
-        transientFeature.setId("transient.feature");
-        transientFeature.setUrl("transient-url");
-        transientFeature.setVersion("1.2.3");
-        transientFeatures.add(transientFeature);
-
-        transientSite.setFeatures(transientFeatures);
-
-        File updatedFile = new File(target, "platform.xml");
-        Platform.write(transientPlatform, updatedFile);
-
-        Platform updated = Platform.read(updatedFile);
-
-        assertEquals("plugins/yyy/,plugins/zzz/", updated.getSites().get(2).getPluginsStr());
-
-        List<Platform.Feature> updatedFeatures = updated.getSites().get(2).getFeatures();
-        assertEquals(1, updatedFeatures.size());
-        assertEquals(transientFeature.getId(), updatedFeatures.get(0).getId());
-        assertEquals(transientFeature.getUrl(), updatedFeatures.get(0).getUrl());
-        assertEquals(transientFeature.getVersion(), updatedFeatures.get(0).getVersion());
     }
 
     @Test